refactor: update header component structure and styling
- Refactor header component exports and imports
- Update HeaderMenu component name and references
- Add responsive styling for header elements
- Fix component naming consistency throughout header module
- Update test references to use new component names
