+++ src/widgets/header/ui/header-calculator/header-calculator.module.scss
@@ -1,13 +1,39 @@
+@use "@/src/app/styles/vendor/variables" as *;
+
+
 .calculator_btn {
 display: flex;
   align-items: center;
   gap: 10px;
   padding: 8px 15px;
+  
+  @media (max-width: $tablet) {
+    padding: 8px 10px;
+    gap: 5px;
+  }
+  
+  @media (max-width: $mobile) {
+    padding: 6px 8px;
+    gap: 3px;
+  }
 }
 
-.image {}
+.image {
+  @media (max-width: $tablet) {
+    width: 32px !important;
+    height: 25px !important;
+  }
+  
+  @media (max-width: $mobile) {
+    width: 28px !important;
+    height: 22px !important;
+  }
+}
 
 .text {
   font-size: 11px;
   font-weight: 700;
+  @media (max-width: 840px){
+    display: none;
+  }
 }
