+++ src/widgets/header/ui/header-factory/header-factory.module.scss
@@ -1,7 +1,12 @@
+@use "@/src/app/styles/vendor/variables" as *;
+
 .factory {
 display: flex;
   gap: 10px;
   align-items: center;
+  @media (max-width: $tablet){
+    display: none;
+  }
 }
 .description {}
