+++ src/widgets/header/ui/header-menu/index.ts
@@ -1,1 +1,1 @@
-export {HeaderSubmenu} from './header-submenu';
+export {HeaderMenu} from './header-menu';
