+++ src/widgets/header/ui/header-messengers/header-messengers.module.scss
@@ -1,9 +1,14 @@
+@use "@/src/app/styles/vendor/variables" as *;
+
 .row {
-display: flex;
+  display: flex;
   gap: 10px;
   justify-content: space-between;
 }
-.messengers {}
+
+.messengers {
+
+}
 
 .title {
   font-size: 11px;
