+++ src/widgets/header/ui/header-submenu/header-submenu.module.scss
@@ -1,0 +1,134 @@
+
+.dropdown {
+  position: absolute;
+  top: 100%;
+  left: 50%;
+  transform: translateX(-50%);
+  margin-top: 2px;
+  animation: dropdownFadeIn 0.3s cubic-bezier(0.4, 0, 0.2, 1);
+  display: none;
+
+  &::after {
+    content: "";
+    position: absolute;
+    top: -10px;
+    left: -20px;
+    right: -20px;
+    height: 12px;
+    background: transparent;
+    pointer-events: auto;
+  }
+
+  &::before {
+    content: "";
+    position: absolute;
+    top: -4px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 0;
+    height: 0;
+    border-left: 6px solid transparent;
+    border-right: 6px solid transparent;
+    border-bottom: 6px solid rgba(40, 40, 40, 0.98);
+    filter: drop-shadow(0 -2px 4px rgba(0, 0, 0, 0.1));
+  }
+}
+
+.dropdownContent {
+  background: rgba(40, 40, 40, 0.98);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  border-radius: 12px;
+  padding: 8px 0;
+  min-width: 200px;
+  backdrop-filter: blur(20px);
+  box-shadow:
+          0 20px 40px rgba(0, 0, 0, 0.3),
+          0 8px 16px rgba(0, 0, 0, 0.2),
+          inset 0 1px 0 rgba(255, 255, 255, 0.1);
+
+  @media (max-width: 768px) {
+    min-width: 180px;
+    border-radius: 8px;
+  }
+
+  @media (max-width: 576px) {
+    min-width: 160px;
+    padding: 6px 0;
+  }
+}
+
+.dropdownItem {
+  display: block;
+  padding: 12px 20px;
+  color: rgba(255, 255, 255, 0.9);
+  text-decoration: none;
+  font-size: 14px;
+  font-weight: 400;
+  letter-spacing: 0.2px;
+  transition: all 0.2s cubic-bezier(0.4, 0, 0.2, 1);
+  border-bottom: 1px solid rgba(255, 255, 255, 0.05);
+  position: relative;
+  overflow: hidden;
+
+  &::before {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: -100%;
+    width: 100%;
+    height: 100%;
+    background: linear-gradient(
+                    90deg,
+                    transparent 0%,
+                    rgba(245, 146, 17, 0.1) 50%,
+                    transparent 100%
+    );
+    transition: left 0.5s ease;
+    pointer-events: none;
+  }
+
+  &:last-child {
+    border-bottom: none;
+  }
+
+  &:hover {
+    background: linear-gradient(135deg, rgba(245, 146, 17, 0.15) 0%, rgba(255, 255, 0, 0.05) 100%);
+    color: #f59211;
+    transform: translateX(4px);
+
+    &::before {
+      left: 100%;
+    }
+  }
+
+  &:active {
+    transform: translateX(2px);
+    background: rgba(245, 146, 17, 0.2);
+  }
+
+  @media (max-width: 768px) {
+    padding: 10px 16px;
+    font-size: 13px;
+  }
+
+  @media (max-width: 576px) {
+    padding: 8px 14px;
+    font-size: 12px;
+  }
+
+  @media (max-width: 480px) {
+    padding: 6px 12px;
+    font-size: 11px;
+  }
+}
+
+@keyframes dropdownFadeIn {
+  0% {
+    opacity: 0;
+    transform: translateX(-50%) translateY(0);
+  }
+  100% {
+    opacity: 1;
+    transform: translateX(-50%) translateY(0);
+  }
+}
