+++ src/widgets/header/ui/header/index.ts
@@ -1,0 +1,1 @@
+export {Header} from './the-header'
