+++ src/widgets/header/ui/header/the-header.tsx
@@ -6,11 +6,11 @@ import {HeaderLogo} from "@/src/widgets/header/ui/header-logo";
 import {HeaderFactory} from "@/src/widgets/header/ui/header-factory";
 import { HeaderCalculator } from '../header-calculator';
 import { HeaderMessengers } from '../header-messengers';
-import {HeaderSubmenu} from "@/src/widgets/header/ui/header-submenu";
-import {HeaderSubmenuItem} from "@/src/widgets/header/ui/header-submenu-item";
 import {NavItem} from "@/src/widgets/header/types";
 import {HeaderActions} from "@/src/widgets/header/ui/header-actions";
 import {ReactNode} from "react";
+import {HeaderMenu} from "@/src/widgets/header/ui";
+import {HeaderMenuItem} from "@/src/widgets/header/ui";
 
 export interface HeaderProps {
 initialNavigationItems: NavItem[];
@@ -37,7 +37,7 @@ export const Header = ({initialNavigationItems,searchButton,cartButton,languageS
                 </div>
                 <div className={css.bottom_header}>
                    <div className="header__container">
-                       <HeaderSubmenu links={navigationItems} Element={HeaderSubmenuItem}/>
+                       <HeaderMenu links={navigationItems} Element={HeaderMenuItem}/>
                    </div>
 
                 </div>
