+++ src/widgets/header/ui/index.ts
@@ -1,0 +1,8 @@
+export {Header} from './header';
+export {HeaderMenu} from './header-menu';
+export {HeaderLogo} from './header-logo';
+export {HeaderMenuItem} from './header-menu-item';
+export {HeaderActions} from './header-actions';
+export {HeaderFactory} from './header-factory'
+export {HeaderCalculator} from './header-calculator';
+export {HeaderMessengers} from './header-messengers';
