+++ src/widgets/index.ts
@@ -1,1 +1,1 @@
-export {Header} from './header'
+export * from './header'
